Fix some warnings & remove fixedstring.*
Some spurious compiler warnings come up when using g++-4.1.2, which are fixed
here for cleanlyness.  The test tools in boost-1.38 also emit some unavoidable
warnings, so changed the way that CMAKE_CXX_FLAGS is set up to allow the user
to easily override -Werror if necessary.

Also removed fixedstring.*, since in hindsight having a static set of standard
named ustrings like this doesn't appear to be particlarly useful.  (What's
more, such static strings can't reliably be used for initializing other static
variables due to static initialization order problems.)

+++ prototypes/newcore/ustring/ustring_test.cpp
@@ -50,7 +50,7 @@ BOOST_AUTO_TEST_CASE(ustring_container_test)
     ustring s("blah");
     BOOST_CHECK_EQUAL(s[0], 'b');
     BOOST_CHECK_EQUAL(s[3], 'h');
-    BOOST_CHECK_EQUAL(s.size(), 4);
+    BOOST_CHECK_EQUAL(s.size(), size_t(4));
     // check iterators
     {
         std::string extract(s.begin(), s.end());
